Fix another bitsize issue related to 081980c
Instead of expanding the specialIntIndex array, cap the bitsizeto 64 before accessing the specialIntIndex. This is because the expanded entries of the specialIntIndex array are useless and cause a crash on use.

In longer words, the `sort.Search` returns `len(specialInts)` if there isn't any entry that satisfies the given function, which is the case when bitsize > 64 given the current `specialInts`. Using the returned value, `len(specialInts)`, to index the `specialInts` array leads to bad behavior.
```go
		specialIntIndex[i] = sort.Search(len(specialInts), func(i int) bool {
			return specialInts[i]>>bitSize != 0
		})
```